requirements.txt
aldjemy==0.10.1
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coverage==5.0.3
Django==3.0.2
django-cors==0.1
django-cors-headers==3.2.1
djangorestframework==3.11.0
ez-setup==0.9
gunicorn==20.0.4
idna==2.8
psycopg2==2.8.4
psycopg2-binary==2.8.4
pytz==2019.3
redis==3.4.1
requests==2.22.0
SQLAlchemy==1.3.13
sqlparse==0.3.0
urllib3==1.25.8
